backend/requirements.txt
djangorestframework~=3.13
Django~=4.0
